+++ examples/io/sys.cpp
@@ -38,7 +38,7 @@ int main() {
 
 	volatile bool quit = 0;
 
-	input.listen_keyboard([&](uint16_t key, fan::keyboard_state state, bool action) {
+	input.listen_keyboard([&](int key, fan::keyboard_state state, bool action) {
 
 		if (state != fan::keyboard_state::press) {
 			return;
