+++ examples/post_process/post_process.cpp
@@ -69,7 +69,7 @@ int main() {
 	pile->loco.set_vsync(false);
 
 	pile->loco.window.add_buttons_callback(pile, 
-		[](fan::window_t* w, uint16_t key, fan::key_state ks, void* userptr) {
+		[](fan::window_t* w, int key, fan::key_state ks, void* userptr) {
 		pile_t* pile = (pile_t*)userptr;
 
 		if (ks != fan::key_state::press) {
